requirements.txt
asgiref==3.8.1
Authlib==1.4.1
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
Flask==3.1.0
Flask-Dance==7.1.0
Flask-OAuthlib==0.9.6
Flask-PyMongo==3.0.1
Flask-SocketIO==5.5.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==2.1.0
packaging==24.2
pycparser==2.22
pymongo==4.11
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
requests-oauthlib==1.1.0
simple-websocket==1.1.0
urllib3==2.3.0
URLObject==2.4.3
Werkzeug==3.1.3
wsproto==1.2.0
